refactor: extract action to connect wallet
